fix(core): re-enable CRA migration to Vite
